requirements.txt
langchain
langchain-community
langchain-groq
langchain_core
uuid
chromadb==0.5.3
pandas
python-dotenv